backend/requirements.txt
functions-framework==3.*
google-generativeai
python-dotenv